+++ app/src/main/java/com/example/bittrack/domain/use_case/GetBtcRateUseCase.kt
@@ -6,6 +6,7 @@ import com.example.bittrack.domain.models.BtcRate
 import com.example.bittrack.domain.repository.RateRepository
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.withContext
+import java.math.BigDecimal
 import javax.inject.Inject
 import kotlin.time.Duration.Companion.hours
 
@@ -15,15 +16,15 @@ class GetBtcRateUseCase @Inject constructor(
 ) {
     private val refreshInterval = 1.hours
 
-    suspend operator fun invoke(): Double = withContext(ioDispatcher) {
+    suspend operator fun invoke(): BigDecimal = withContext(ioDispatcher) {
         val now = DateUtils.nowMillis()
         val localRate = rateRepository.getCachedBtcUsdRate()
         if (localRate == null || now - localRate.timestamp >= refreshInterval.inWholeMilliseconds) {
             val remoteRate = rateRepository.fetchRemoteBtcUsdRate()
             remoteRate?.let { rateRepository.cacheBtcUsdRate(BtcRate(it, now)) }
-            return@withContext remoteRate ?: 0.0
+            return@withContext BigDecimal(remoteRate ?: 0.0)
         } else {
-            return@withContext localRate.rate
+            return@withContext BigDecimal(localRate.rate)
         }
     }
 }
